// libs/interface/service/service.details.interface.ts
import { MDAInterface } from "../mda/mdas.interface";
import {
  RelatedServices,
  ServiceApplicationSteps,
  ServiceDocuments,
  ServiceFAQs,
  ServiceFees,
  ServiceLocations,
} from "./service.helpers";

export interface ServiceDetailsInterface {
  id: string;
  name?: string;
  page?: string;
  description?: string;
  processing_time?: string;
  price?: string;
  service_provider?: string;
  availability?: string;
  online_application_url?: string;
  verified: boolean;
  ministry_id?: string;
  ministry?: Pick<MDAInterface, "id" | "name" | "contact"> | null;
  who_can_apply?: [];
  eligibility_requirements?: [];
  documents_required?: ServiceDocuments[];
  application_steps?: ServiceApplicationSteps[];
  fees?: ServiceFees[];
  locations?: ServiceLocations[];
  faqs?: ServiceFAQs[];
  related_services?: RelatedServices[];
  important_notes?: [];
  has_detail_page: boolean;
  active: boolean;
}